+++ CHANGELOG.md
@@ -42,6 +42,7 @@ high state of flux, you're at risk of it changing without notice.
     - `WriterT` (@gcanti)
     - `ValidationT` (@gcanti)
   - **Modules**
+    - rename `ReadonlyArray` module to `Array` (@gcanti)
     - `Applicative`
       - remove `getApplicativeComposition` and related types (@gcanti)
     - `Compactable`